Refactoring; removing unneeded usecase; chaning User entitiy so that ID is String
