--- src/interface_adapter/clear_users/ClearController.java
@@ -1,14 +1,0 @@
-package interface_adapter.clear_users;
-
-import use_case.clear_users.ClearInputBoundary;
-
-public class ClearController {
-
-    final ClearInputBoundary clearUseCaseInteractor;
-
-    public ClearController(ClearInputBoundary clearUseCaseInteractor){
-        this.clearUseCaseInteractor = clearUseCaseInteractor;
-    }
-
-    public void execute(){ clearUseCaseInteractor.execute();}
-}
