--- src/use_case/clear_users/ClearInputBoundary.java
@@ -1,7 +1,0 @@
-package use_case.clear_users;
-
-
-public interface ClearInputBoundary {
-    void execute();
-    //Remove all users from the database
-}
