--- src/use_case/clear_users/ClearInputData.java
@@ -1,6 +1,0 @@
-package use_case.clear_users;
-
-// I decided not to implement this class. I think it is not necessary.
-
-public class ClearInputData {
-}
