--- src/use_case/clear_users/ClearOutputBoundary.java
@@ -1,8 +1,0 @@
-package use_case.clear_users;
-
-
-public interface ClearOutputBoundary {
-    void prepareSuccessView(ClearOutputData clearOutputData);
-
-    void prepareFailView(String error);
-}
