--- src/use_case/clear_users/ClearOutputData.java
@@ -1,17 +1,0 @@
-package use_case.clear_users;
-
-public class ClearOutputData {
-    private final boolean useCaseFailed;
-
-    private final String[] usersCleared;
-
-    public ClearOutputData(String[] usersCleared, boolean useCaseFailed) {
-
-        this.useCaseFailed = useCaseFailed;
-        this.usersCleared = usersCleared;
-    }
-
-    public String[] getUsersCleared() {
-        return usersCleared;
-    }
-}
