+++ src/use_case/menu/view/MenuView.java
@@ -2,8 +2,8 @@ package view;
 
 import interface_adapter.Menu.MenuController;
 import interface_adapter.Menu.MenuViewModel;
-import interface_adapter.leaderboard.LeaderboardController;
-import interface_adapter.leaderboard.LeaderboardViewModel;
+import use_case.leaderboard.interface_adapter.LeaderboardController;
+import use_case.leaderboard.interface_adapter.LeaderboardViewModel;
 import interface_adapter.signup.SignupState;
 import interface_adapter.signup.SignupViewModel;
