--- users.csv
@@ -1,2 +1,0 @@
-username,password,user_id
-Peter,11,1
